Move the provider search into its own translation unit
- moved the candidate model and search types to craft_reservation_search.h
- moved exclusion signatures, candidate classes, fingerprint and search to its cpp
- moved the pseudo-item and candidate identity helpers with them
- added selection_context::pool_feasible for the charged-pool prune
- kept discovery, validation and the debit in crafting.cpp
